backend/requirements.txt
asgiref==3.5.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
coreapi==2.3.3
coreschema==0.0.4
cryptography==36.0.2
defusedxml==0.7.1
Django==4.0.3
django-allauth==0.49.0
django-bootstrap-modal-forms==2.2.0
django-cors-headers==3.7.0
django-crispy-forms==1.14.0
django-cron==0.6.0
django-filter==2.4.0
django-templated-mail==1.1.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
exceptiongroup==1.1.3
factory-boy==3.3.0
flake8==3.8.3
idna==3.3
iniconfig==2.0.0
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.3
mysqlclient==2.2.0
mccabe==0.6.1
oauthlib==3.2.0
packaging==23.1
pluggy==1.3.0
psycopg2==2.9.3
pycodestyle==2.6.0
pycparser==2.21
pyflakes==2.2.0
PyJWT==2.3.0
pytest==7.4.0
pytest-django==4.5.2
python3-openid==3.2.0
pytz==2022.1
requests==2.27.1
requests-oauthlib==1.3.1
scikit-learn==1.3.2
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.4.2
sqlparse==0.4.2
tomli==2.0.1
tzdata==2022.1
uritemplate==4.1.1
urllib3==1.26.9
